- Shorten import statement to ``from qrbill import QRBill``. - Add Github Actions to test the code. - Refactored all const vars to a separate file called constants.py. - Put the most important class QRBill at the top of the file. - Enforced a 80 column on the source code. - Add requirements.txt and fixing dependency versions.
